myproject/requirements.txt
anyio==4.0.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
beautifulsoup4==4.12.2
binaryornot==0.4.4
certifi==2023.7.22
channels==4.0.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cookiecutter==2.3.0
cssbeautifier==1.14.9
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-admin-sortable2==2.1.9
django-browser-reload==1.11.0
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-grappelli==3.0.7
django-import-export==3.3.1
django-js-asset==2.1.0
django-mptt==0.14.0
django-render-block==0.9.2
django-taggit==4.0.0
django-tailwind==3.6.0
django-widget-tweaks==1.5.0
djlint==1.32.1
EditorConfig==0.12.3
et-xmlfile==1.1.0
executing==1.2.0
gunicorn==21.2.0
h11==0.14.0
html-tag-names==0.1.2
html-void-elements==0.1.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
jsbeautifier==1.14.9
json5==0.9.14
markdown-it-py==3.0.0
MarkupPy==1.14
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.39
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
python-slugify==8.0.1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rich==13.5.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
tablib==3.5.0
text-unidecode==1.3
tqdm==4.66.1
traitlets==5.10.0
tzdata==2023.3
urllib3==2.0.4
uwsgi==2.0.23
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
